Move `safeNumber` to shared util and tweak implementation
