+++ packages/loot-core/src/server/budget/util.js
@@ -1,11 +1,14 @@
 import { number } from '../spreadsheet/globals';
+import { safeNumber } from '../../shared/util';
 
 export { number } from '../spreadsheet/globals';
 
 export function sumAmounts(...amounts) {
-  return amounts.reduce((total, amount) => {
-    return total + number(amount);
-  }, 0);
+  return safeNumber(
+    amounts.reduce((total, amount) => {
+      return total + number(amount);
+    }, 0)
+  );
 }
 
 export function flatten2(arr) {
@@ -18,24 +21,4 @@ export function unflatten2(arr) {
     res.push([arr[i], arr[i + 1]]);
   }
   return res;
-}
-
-// Note that we don't restrict values to `Number.MIN_SAFE_INTEGER <= value <= Number.MAX_SAFE_INTEGER`
-// where `Number.MAX_SAFE_INTEGER == 2^53 - 1` but a smaller range over `-(2^43-1) <= value <= 2^43 - 1`.
-// This ensure that the number is accurate not just for the integer component but for 3 decimal places also.
-//
-// This gives us the guarantee that can use `safeNumber` on number whether they are unscaled user inputs
-// or they have been converted to integers (using `amountToInteger`).
-
-const MAX_SAFE_NUMBER = 2 ** 43 - 1;
-const MIN_SAFE_NUMBER = -MAX_SAFE_NUMBER;
-
-export function safeNumber(value) {
-  value = number(value);
-  if (value > MAX_SAFE_NUMBER || value < MIN_SAFE_NUMBER) {
-    throw new Error(
-      "Can't safely perform arithmetic with number: " + JSON.stringify(value)
-    );
-  }
-  return value;
 }
