+++ packages/loot-core/src/shared/util.js
@@ -298,6 +298,30 @@ export function getNumberFormat() {
 
 setNumberFormat('comma-dot');
 
+// Number utilities
+
+// We dont use `Number.MAX_SAFE_NUMBER` and such here because those
+// numbers are so large that it's not safe to convert them to floats
+// (i.e. N / 100). For example, `9007199254740987 / 100 ===
+// 90071992547409.88`. While the internal arithemetic would be correct
+// because we always do that on numbers, the app would potentially
+// display wrong numbers. Instead of `2**53` we use `2**51` which
+// gives division more room to be correct
+const MAX_SAFE_NUMBER = 2**51 - 1;
+const MIN_SAFE_NUMBER = -MAX_SAFE_NUMBER;
+
+export function safeNumber(value) {
+  if (!Number.isInteger(value)) {
+    throw new Error('safeNumber: number is not an integer: ' + value);
+  }
+  if (value > MAX_SAFE_NUMBER || value < MIN_SAFE_NUMBER) {
+    throw new Error(
+      "safeNumber: can't safely perform arithmetic with number: " + value
+    );
+  }
+  return value;
+}
+
 export function toRelaxedNumber(value) {
   return integerToAmount(currencyToInteger(value) || 0);
 }
@@ -307,8 +331,7 @@ export function toRelaxedInteger(value) {
 }
 
 export function integerToCurrency(n) {
-  // Awesome
-  return numberFormat.formatter.format(n / 100);
+  return numberFormat.formatter.format(safeNumber(n) / 100);
 }
 
 export function amountToCurrency(n) {
@@ -340,7 +363,7 @@ export function amountToInteger(n) {
 }
 
 export function integerToAmount(n) {
-  return parseFloat((n / 100).toFixed(2));
+  return parseFloat((safeNumber(n) / 100).toFixed(2));
 }
 
 // This is used when the input format could be anything (from
